Исправления по замечаниям ревю
