+++ Tracker/Cells/ScheduleCell.swift
@@ -52,13 +52,13 @@ final class ScheduleCell: UITableViewCell {
         if switcher.isOn {
             selectedDays.append(title.text ?? "")
             switch title.text {
-            case dayOfWeek.monday.rawValue: shortSelectedDays.append("ПН")
-            case dayOfWeek.tuesday.rawValue: shortSelectedDays.append("ВТ")
-            case dayOfWeek.wednesday.rawValue: shortSelectedDays.append("СР")
-            case dayOfWeek.thursday.rawValue: shortSelectedDays.append("ЧТ")
-            case dayOfWeek.friday.rawValue: shortSelectedDays.append("ПТ")
-            case dayOfWeek.saturday.rawValue: shortSelectedDays.append("СБ")
-            case dayOfWeek.sunday.rawValue: shortSelectedDays.append("ВС")
+            case weekdays.monday.rawValue: shortSelectedDays.append("ПН")
+            case weekdays.tuesday.rawValue: shortSelectedDays.append("ВТ")
+            case weekdays.wednesday.rawValue: shortSelectedDays.append("СР")
+            case weekdays.thursday.rawValue: shortSelectedDays.append("ЧТ")
+            case weekdays.friday.rawValue: shortSelectedDays.append("ПТ")
+            case weekdays.saturday.rawValue: shortSelectedDays.append("СБ")
+            case weekdays.sunday.rawValue: shortSelectedDays.append("ВС")
             case .none:
                 return
             case .some(_):
