+++ Tracker/Helpers/UIColor+Extensions.swift
@@ -1,4 +1,3 @@
-//
 //  UIColor+Extensions.swift
 //  Tracker
 //
@@ -8,12 +7,22 @@
 import UIKit
 
 extension UIColor {
-    static var YPGray: UIColor { UIColor(named: "YPGray") ?? UIColor.gray }
-    static var YPRed: UIColor { UIColor(named: "YPRed") ?? UIColor.red }
-    static var YPBlack: UIColor { UIColor(named: "YPBlack") ?? UIColor.black }
-    static var YPBlue: UIColor { UIColor(named: "YPBlue") ?? UIColor.blue }
-    static var BackgroundColor: UIColor { UIColor(named: "BackgroundColor") ?? UIColor.red }
-    static var YPWhite: UIColor { UIColor(named: "YPWhite") ?? UIColor.white }
-    static var YPLightGray: UIColor { UIColor(named: "YPLightGray") ?? UIColor.lightGray}
-    
+    static var Color_one: UIColor { UIColor(named: "Color_one") ?? UIColor.red }
+    static var Color_two: UIColor { UIColor(named: "Color_two") ?? UIColor.red }
+    static var Color_three: UIColor { UIColor(named: "Color_three") ?? UIColor.red }
+    static var Color_four: UIColor { UIColor(named: "Color_four") ?? UIColor.red }
+    static var Color_five: UIColor { UIColor(named: "Color_five") ?? UIColor.red }
+    static var Color_six: UIColor { UIColor(named: "Color_six") ?? UIColor.red }
+    static var Color_seven: UIColor { UIColor(named: "Color_seven") ?? UIColor.red }
+    static var Color_eight: UIColor { UIColor(named: "Color_eight") ?? UIColor.red }
+    static var Color_nine: UIColor { UIColor(named: "Color_nine") ?? UIColor.red }
+    static var Color_ten: UIColor { UIColor(named: "Color_ten") ?? UIColor.red }
+    static var Color_eleven: UIColor { UIColor(named: "Color_eleven") ?? UIColor.red }
+    static var Color_twelve: UIColor { UIColor(named: "Color_twelve") ?? UIColor.red }
+    static var Color_thirteen: UIColor { UIColor(named: "Color_thirteen") ?? UIColor.red }
+    static var Сolor_fourteen: UIColor { UIColor(named: "Сolor_fourteen") ?? UIColor.red }
+    static var Сolor_fiveteen: UIColor { UIColor(named: "Сolor_fiveteen") ?? UIColor.red }
+    static var Color_sixteen: UIColor { UIColor(named: "Color_sixteen") ?? UIColor.red }
+    static var Color_seventeen: UIColor { UIColor(named: "Color_seventeen") ?? UIColor.red }
+    static var Color_eighteen: UIColor { UIColor(named: "Color_eighteen") ?? UIColor.red }
 }
