+++ Tracker/TabBar/TrackersViewController.swift
@@ -1,16 +1,11 @@
 //
-//  TrackersViewControllerProtocol.swift
+//  TrackersViewController.swift
 //  Tracker
 //
 //  Created by Алексей Налимов on 08.10.2023.
 //
 
 import UIKit
-
-protocol TrackersViewControllerProtocol {
-    func saveDoneEvent(id: UUID, index: IndexPath)
-    var localTrackers: [TrackerCategory] {get set }
-}
 
 class TrackersViewController: UIViewController {
 
@@ -194,7 +189,7 @@ class TrackersViewController: UIViewController {
 
     @objc
     private func plusTapped() {
-        let selecterTrackerVC = TrackerTypeViewController()
+        let selecterTrackerVC = TypeViewController()
         show(selecterTrackerVC, sender: self)
     }
